backend/requirements/base.txt
# ============================================
# DEPENDENCIAS BASE - COMUNES A TODOS LOS ENTORNOS
# ============================================

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================
# VALIDATION AND SERIALIZATION
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
typing-extensions==4.8.0
annotated-types==0.6.0

# ============================================
# SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# ============================================
# DATE AND TIME
# ============================================
python-dateutil==2.8.2
pytz==2023.3

# ============================================
# HTTP CLIENT
# ============================================
httpx==0.25.2
requests==2.31.0

# ============================================
# FILE PROCESSING
# ============================================
openpyxl==3.1.2
pandas==2.1.3
numpy==1.26.2

# ============================================
# NOTIFICATIONS
# ============================================
twilio==8.2.0
aiosmtplib==3.0.1
jinja2==3.1.2

# ============================================
# TASK SCHEDULING
# ============================================
apscheduler==3.10.4

# ============================================
# MONITORING AND LOGGING
# ============================================
python-json-logger==2.0.7
